requirements.txt
requests
PyExecjs
PyInstaller
uncompyle6
